refactor: Move selector to appropriate dir
