+++ src/components/transactions/TransactionFormModal.svelte
@@ -8,7 +8,7 @@ import { toastStore } from '../../stores/toast.svelte';
 import { reactiveTransactions } from '../../stores/transactions.svelte';
 import BottomSheet from '../ui/BottomSheet.svelte';
 import Button from '../ui/Button.svelte';
-import CategorySelector from '../ui/CategorySelector.svelte';
+import CategorySelector from './CategorySelector.svelte';
 import Icon from '../ui/Icon.svelte';
 import Input from '../ui/Input.svelte';
